(Qt) Combine qt/options/generic.cpp and qt/options/nongeneric.cpp
